Implements a string_view formatter parser.
This wraps libc++'s std::formatter for a std::string_view and uses its own
parser to validate the std-format-spec. This makes it possible to use our
improved diagnostics when parsing the std-format-spec of standard
formatters.

The code heavily replies on the implementation details of libc++'s
formatter; it has to provide the same parsed state as the formatter's own
parse function.

This method can be used for formatter specializations of user-defined
types.