Improves the fixed_string class.
Make the class more looke like a real string class instead of always
looking at the internals. Note due to language requirements private
mambers can't be used.

Allow wchar_t specializations of the class to be created, these are not
used at the moment but will be in the future.

As a side effect change more calls to format_error's constructor to use
the crete_format_error helper function.